requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
aws-cdk-lib>=2.1.0, <3.0.0
constructs>=10.0.5, <11.0.0
cdk_nag